pyproject.toml
[project]
name = "childbook-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.3.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
